audio-visual-script-repo/python/requirements.txt
# Core Python dependencies will be added as modules land.
pyzmq
requests
pytest
